+++ nut/app/events/events.c
@@ -120,7 +120,7 @@ THREAD(Low, arg)
  */
 int main(void)
 {
-    u_long baud = 115200;
+    uint32_t baud = 115200;
 
     /*
      * Register the UART device, open it, assign stdout to it and set
